+++ src/youwol/backends/accounts/openid_rp/openid_flows_service.py
@@ -7,18 +7,14 @@ from typing import Callable, Optional, Tuple
 # Youwol utilities
 from youwol.utils import CacheClient
 from youwol.utils.clients.oidc.oidc_config import OidcForClient
-from youwol.utils.clients.oidc.tokens import (
-    Tokens,
-    restore_tokens_from_session_id,
-    save_tokens,
-)
+from youwol.utils.clients.oidc.tokens_manager import Tokens, TokensManager
 
 # relative
 from .openid_flows_states import AuthorizationFlow, LogoutFlow
 
 
 class FlowStateNotFound(RuntimeError):
-    def __init__(self):
+    def __init__(self) -> None:
         super().__init__("Flow state not found")
 
 
@@ -31,6 +27,9 @@ class OpenidFlowsService:
     ):
         self.__cache = cache
         self.__oidc_client = oidc_client
+        self.__tokens_manager = TokensManager(
+            cache=self.__cache, oidc_client=self.__oidc_client
+        )
         self.__tokens_id_generator = tokens_id_generator
 
     async def init_authorization_flow(
@@ -70,11 +69,9 @@ class OpenidFlowsService:
             code=code, code_verifier=code_verifier, redirect_uri=callback_uri
         )
 
-        tokens = await save_tokens(
+        tokens = await self.__tokens_manager.save_tokens(
             tokens_id=self.__tokens_id_generator(),
-            cache=self.__cache,
-            oidc_client=self.__oidc_client,
-            **tokens_data,
+            tokens_data=tokens_data,
         )
 
         return tokens, target_uri
@@ -84,16 +81,14 @@ class OpenidFlowsService:
             username=username, password=password
         )
 
-        return await save_tokens(
+        return await self.__tokens_manager.save_tokens(
             tokens_id=self.__tokens_id_generator(),
-            cache=self.__cache,
-            oidc_client=self.__oidc_client,
-            **tokens_data,
+            tokens_data=tokens_data,
         )
 
     async def init_logout_flow(
         self, target_uri: str, forget_me: bool, callback_uri: str
-    ):
+    ) -> str:
         logout_flow_state = LogoutFlow(
             uuid=str(uuid.uuid4()),
             target_uri=target_uri,
@@ -108,7 +103,7 @@ class OpenidFlowsService:
 
         return url
 
-    def handle_logout_flow_callback(self, flow_uuid):
+    def handle_logout_flow_callback(self, flow_uuid: str) -> Tuple[str, bool]:
         flow_state_data = self.__cache.get(LogoutFlow.cache_key(flow_uuid))
 
         if flow_state_data is None:
@@ -121,12 +116,14 @@ class OpenidFlowsService:
 
         return target_uri, forget_me
 
-    async def handle_logout_back_channel(self, logout_token: str):
+    async def handle_logout_back_channel(self, logout_token: str) -> None:
         logout_token_decoded = await self.__oidc_client.token_decode(logout_token)
         # TODO : validate logout token (see https://openid.net/specs/openid-connect-backchannel-1_0.html#Validation)
-        tokens = restore_tokens_from_session_id(
+        tokens = self.__tokens_manager.restore_tokens_from_session_id(
             session_id=logout_token_decoded["sid"],
-            cache=self.__cache,
-            oidc_client=self.__oidc_client,
         )
+
+        if tokens is None:
+            raise FlowStateNotFound()
+
         await tokens.delete()
